+++ src/main/java/so/alaz/cleansingpotions/CleansingPotionsFabricClient.java
@@ -2,10 +2,12 @@ package so.alaz.cleansingpotions;
 
 //? fabric {
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.minecraft.client.renderer.entity.NoopRenderer;
 import net.minecraft.client.renderer.entity.ThrownItemRenderer;
 import so.alaz.cleansingpotions.entity.ModEntities;
+import so.alaz.cleansingpotions.update.UpdateNotifier;
 
 public class CleansingPotionsFabricClient implements ClientModInitializer {
 
@@ -13,6 +15,7 @@ public class CleansingPotionsFabricClient implements ClientModInitializer {
     public void onInitializeClient() {
         EntityRendererRegistry.register(ModEntities.CLEANSING_POTION, ThrownItemRenderer::new);
         EntityRendererRegistry.register(ModEntities.CLEANSING_CLOUD, NoopRenderer::new);
+        ClientPlayConnectionEvents.JOIN.register((handler, sender, client) -> UpdateNotifier.onJoin());
     }
 }
 //?}